separate klayout_tech from LayerDisplayProperties
